requirements.txt
python-telegram-bot==20.7
openai
python-dotenv
psycopg2-binary
